chien2

+++ DataModel/ThongKeKhachModel.cs
@@ -11,8 +11,7 @@ namespace DataModel
         public int MaSach { get; set; }
         public string TenSach { get; set; }
         public int SoLuong { get; set; }
-        public Decimal Giaban { get; set; }
-        //public Decimal TongGia { get; set; }
+        public int Giaban { get; set; }
         public DateTime NgayLapHD { get; set; }
         public string TenKH { get; set; }
         public string Diachi { get; set; }
